python/examples/custom-threat-service/requirements.txt
coverage==3.7.1
Django==1.8.2
djangorestframework==3.1.3
funcsigs==0.4
futures==3.0.3
mock==1.1.3
pbr==1.3.0
pep8==1.6.2
six==1.9.0
wheel==0.24.0
